Reduce the size of the inset, pad and color structs.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@30514 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/ewl/src/lib/ewl_object.h
@@ -83,10 +83,10 @@ struct Ewl_Object
 
 	struct
 	{
-		int l, /**< Left value */
-		    r, /**< Right value */
-		    t, /**< Top value */
-		    b; /**< Bottom value */
+		short l, /**< Left value */
+		      r, /**< Right value */
+		      t, /**< Top value */
+		      b; /**< Bottom value */
 	} pad, /**< The space padded around the outside of the object. */
 	insets; /**< The space inside where children should not be laid out. */
